Promote feedback model to a full AR model for persistence
In #421 we moved email sending to custom jobs so that we could provide
better error handling. However, the feedback email wasn't done since
that was based on a custom Active Model model. To work around this
change the model to an Active Record model so that it is persisted to
the database and Active Job can serialize it via Global ID.

This has the added benefit of storing the feedback should the email sent
to the feedback address somehow get lost due to spam filtering, etc.